requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-genai==1.10.0
Pillow==11.1.0
python-multipart==0.0.9
pydantic==2.7.1